Homepage court search (#4)
* Added Court Search Functionality

Home page supports searching for Courts by munipality.  Created a page
to display the search results.

* Clean up a add Print Button

Removed some commented code and added print button to search results
page

* Trying to get leaflet to work

Leaflet is working but when I try to use a marker, it inserts a
bower_components path in front of the raw base64 encoded data.

* Removed extraneous text

Removed Map Goes Here Text

* Icon Update and Fixes for Court Search Map

Added custom markers to the map and changed the leaflet bower version to
0.7.0 since 1.0.1 was having issues with the marker icon path

* Print Button Fix

Removed the print button from the print page

* Home Page Court Search Pull Request Edits

re-did the controller so that it searches off of the url for the id of
the court.  Updated the controller name and removed commented out code

* Handle invalid Court Id

used the error portion of the promise to send a toaster message if the
court Id was invalid

* Toast message for court not found

Added Toast message for and invalid court ID

* Court Search Restructuring

Changed court search URL path and removed courtId being passed as a
parameter.
Also, changed the searching of the courtId so that it occurs in the
resolve function.